create genericform component
